Move NativeImageVM and its dependencies to mx_sdk_benchmark
